+++ ios/Classes/FcmChannelsManagerPlugin.m
@@ -1,20 +1,15 @@
 #import "FcmChannelsManagerPlugin.h"
+#if __has_include(<fcm_channels_manager/fcm_channels_manager-Swift.h>)
+#import <fcm_channels_manager/fcm_channels_manager-Swift.h>
+#else
+// Support project import fallback if the generated compatibility header
+// is not copied when this plugin is created as a library.
+// https://forums.swift.org/t/swift-static-libraries-dont-copy-generated-objective-c-header/19816
+#import "fcm_channels_manager-Swift.h"
+#endif
 
 @implementation FcmChannelsManagerPlugin
 + (void)registerWithRegistrar:(NSObject<FlutterPluginRegistrar>*)registrar {
-  FlutterMethodChannel* channel = [FlutterMethodChannel
-      methodChannelWithName:@"fcm_channels_manager"
-            binaryMessenger:[registrar messenger]];
-  FcmChannelsManagerPlugin* instance = [[FcmChannelsManagerPlugin alloc] init];
-  [registrar addMethodCallDelegate:instance channel:channel];
+  [SwiftFcmChannelsManagerPlugin registerWithRegistrar:registrar];
 }
-
-- (void)handleMethodCall:(FlutterMethodCall*)call result:(FlutterResult)result {
-  if ([@"getPlatformVersion" isEqualToString:call.method]) {
-    result([@"iOS " stringByAppendingString:[[UIDevice currentDevice] systemVersion]]);
-  } else {
-    result(FlutterMethodNotImplemented);
-  }
-}
-
 @end