+++ lib/fcm_channels_manager_method_channel.dart
@@ -35,4 +35,19 @@ class MethodChannelFcmChannelsManager extends FcmChannelsManagerPlatform {
     }
     return channels;
   }
+
+  @override
+  Future<String?> requestNotificationPermission(
+      Map<String, bool> params) async {
+    final res = await methodChannel.invokeMethod<String>(
+        'requestNotificationPermission', params);
+    return res;
+  }
+
+  @override
+  Future<String?> getNotificationPermissionStatus() async {
+    final res = await methodChannel
+        .invokeMethod<String>('getNotificationPermissionStatus');
+    return res;
+  }
 }