+++ CHANGELOG.md
@@ -13,3 +13,9 @@
 ## 0.0.4
 
 * Update Example
+
+## 0.0.5
+
+Improvent request notification permission for iOS
+* New `requestNotificationPermission` function support full options for iOS
+* Get current permission status with `getNotificationStatus`